- finally managed to get rid of the DWORD type.
This one was particularly nasty because Windows also defines a DWORD, but in Windows it is an unsigned long, not an unsigned int so changing types caused type conflicts and not all could be removed.
Those referring to the Windows type have to be kept, fortunately they are mostly in the Win32 directory, with a handful of exceptions elsewhere.

# Conflicts:
#	src/gl/system/gl_swframebuffer.cpp
#	src/gl/system/gl_swwipe.cpp
#	src/gl/utility/gl_clock.cpp
#	src/win32/win32gliface.cpp