- added a modified version of MBF's stay-on-lift feature.
The reason this was never added was the hard dependency on the line trigger types. This implements some modified logic that does not try to find all potential lifts in the map.
Also moving the MBF flags to compatflags so that they are easier to control by the user as these must be part of compatibility presets.

# Conflicts:
#	src/d_main.cpp
#	src/doomdef.h
#	src/g_mapinfo.cpp
#	src/gamedata/g_mapinfo.h
#	src/maploader/specials.cpp
#	src/p_enemy.cpp
#	src/r_defs.h